+++ src/Api/Subscriptions.php
@@ -11,10 +11,10 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Stripe
- * @version    2.0.5
+ * @version    2.0.8
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
- * @copyright  (c) 2011-2016, Cartalyst LLC
+ * @copyright  (c) 2011-2017, Cartalyst LLC
  * @link       http://cartalyst.com
  */
 
@@ -129,6 +129,8 @@ class Subscriptions extends Api
      */
     public function all($customerId, array $parameters = [])
     {
-        return $this->_get("customers/{$customerId}/subscriptions", $parameters);
+        return $this->_get('subscriptions', array_merge($parameters, [
+            'customer' => $customerId
+        ]));
     }
 }