+++ src/Config.php
@@ -11,10 +11,10 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Stripe
- * @version    2.0.5
+ * @version    2.0.8
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
- * @copyright  (c) 2011-2016, Cartalyst LLC
+ * @copyright  (c) 2011-2017, Cartalyst LLC
  * @link       http://cartalyst.com
  */
 
@@ -51,6 +51,13 @@ class Config implements ConfigInterface
     protected $idempotencyKey;
 
     /**
+     * The managed account id.
+     *
+     * @var string
+     */
+    protected $accountId;
+
+    /**
      * Constructor.
      *
      * @param  string  $version
@@ -65,7 +72,7 @@ class Config implements ConfigInterface
 
         $this->setApiKey($apiKey ?: getenv('STRIPE_API_KEY'));
 
-        $this->setApiVersion($apiVersion ?: getenv('STRIPE_API_VERSION') ?: '2015-03-24');
+        $this->setApiVersion($apiVersion ?: getenv('STRIPE_API_VERSION') ?: '2016-07-06');
 
         if (! $this->apiKey) {
             throw new \RuntimeException('The Stripe API key is not defined!');
@@ -140,6 +147,29 @@ class Config implements ConfigInterface
     public function setIdempotencyKey($idempotencyKey)
     {
         $this->idempotencyKey = $idempotencyKey;
+
+        return $this;
+    }
+
+    /**
+     * Returns the managed account id.
+     *
+     * @return string
+     */
+    public function getAccountId()
+    {
+        return $this->accountId;
+    }
+
+    /**
+     * Sets the managed account id.
+     *
+     * @param  string  $accountId
+     * @return $this
+     */
+    public function setAccountId($accountId)
+    {
+        $this->accountId = $accountId;
 
         return $this;
     }