+++ src/Exception/Handler.php
@@ -11,10 +11,10 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Stripe
- * @version    2.0.5
+ * @version    2.0.8
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
- * @copyright  (c) 2011-2016, Cartalyst LLC
+ * @copyright  (c) 2011-2017, Cartalyst LLC
  * @link       http://cartalyst.com
  */
 
@@ -76,7 +76,9 @@ class Handler
 
         $statusCode = $response->getStatusCode();
 
-        $error = json_decode($response->getBody(true), true)['error'];
+        $rawOutput = json_decode($response->getBody(true), true);
+
+        $error = isset($rawOutput['error']) ? $rawOutput['error'] : [];
 
         $errorCode = isset($error['code']) ? $error['code'] : null;
 
@@ -87,7 +89,7 @@ class Handler
         $missingParameter = isset($error['param']) ? $error['param'] : null;
 
         $this->handleException(
-            $message, $statusCode, $errorType, $errorCode, $missingParameter
+            $message, $statusCode, $errorType, $errorCode, $missingParameter, $rawOutput
         );
     }
 
@@ -102,7 +104,7 @@ class Handler
      * @return void
      * @throws \Cartalyst\Stripe\Exception\StripeException
      */
-    protected function handleException($message, $statusCode, $errorType, $errorCode, $missingParameter)
+    protected function handleException($message, $statusCode, $errorType, $errorCode, $missingParameter, $rawOutput)
     {
         if ($statusCode === 400 && $errorCode === 'rate_limit') {
             $class = 'ApiLimitExceeded';
@@ -123,6 +125,7 @@ class Handler
         $instance->setErrorCode($errorCode);
         $instance->setErrorType($errorType);
         $instance->setMissingParameter($missingParameter);
+        $instance->setRawOutput($rawOutput);
 
         throw $instance;
     }