+++ src/Pager.php
@@ -11,10 +11,10 @@
  * bundled with this package in the LICENSE file.
  *
  * @package    Stripe
- * @version    2.0.5
+ * @version    2.0.8
  * @author     Cartalyst LLC
  * @license    BSD License (3-clause)
- * @copyright  (c) 2011-2016, Cartalyst LLC
+ * @copyright  (c) 2011-2017, Cartalyst LLC
  * @link       http://cartalyst.com
  */
 
@@ -84,6 +84,12 @@ class Pager
             $id = $parameters[0];
 
             unset($parameters[0]);
+
+            if (isset($parameters[1])) {
+                $parameters = $parameters[1];
+
+                unset($parameters[1]);
+            }
 
             $parameters = [ $id, $parameters ];
         } else {